REFACTOR - extract postSummary to function in util
