+++ writeSummaryToCSV.js
@@ -1,71 +1,14 @@
 const path = require('path')
-const { escapeComment, stripHTML, writeHeader, appendRow, 
-  getWordCount, median } = require('./util') // Adjust the path as necessary
+const { escapeComment, stripHTML, writeHeader, appendRow, postStatistics } = require('./util') // Adjust the path as necessary
 
-
-// Function to calculate the topic summary
 const topicSummary = (topic) => {
   const posts = topic.replies.flat()
   const topicCreatedAt = new Date(topic.topicCreatedAt)
 
-  // Number of posts
-  const numberOfPosts = posts.length
+  const postSummary = postStatistics(posts, topicCreatedAt)
 
-  if (numberOfPosts === 0) {
-    return {
-      numberOfPosts: 0,
-      medianWordCount: 0,
-      averageTimeDiff: 0,
-      firstReplyTimestamp: 0,
-      averageTimeToPostFromFirst: 0,
-      averagePostsPerAuthor: 0
-    }
-  }
+  return postSummary
 
-
-  // Word counts
-  const wordCounts = posts.map(post => getWordCount(post.postMessage))
-
-  const medianWordCount = Math.round(median(wordCounts) * 10) / 10
-
-  // Average time in hours from topicCreatedAt to postTimestamp
-  const timeDiffs = posts.map(post => {
-    const postTimestamp = new Date(post.postTimestamp)
-    return (postTimestamp - topicCreatedAt) / (1000 * 60 * 60) // Convert from milliseconds to hours
-  })
-
-  const averageTimeDiff = Math.round((timeDiffs.reduce((acc, curr) => acc + curr, 0) / timeDiffs.length) * 10) / 10
-
-  // Average number of posts per postAuthorId
-  const postCountsByAuthor = posts.reduce((acc, post) => {
-    acc[post.postAuthorId] = (acc[post.postAuthorId] || 0) + 1
-    return acc
-  }, {})
-  const averagePostsPerAuthor = Math.round((Object.values(postCountsByAuthor).reduce((acc, curr) => acc + curr, 0) / Object.keys(postCountsByAuthor).length) * 10) / 10
-
-  
-  const firstReplyTimestamp = new Date(Math.min(...posts.map(post => new Date(post.postTimestamp))))
-  const timeDiffsFromFirst = posts
-    .map(post => {
-      const postTimestamp = new Date(post.postTimestamp)
-      return postTimestamp > firstReplyTimestamp 
-        ? (postTimestamp - firstReplyTimestamp) / (1000 * 60 * 60) // Convert from milliseconds to hours
-        : null
-    })
-    .filter(diff => diff !== null)
-  
-  const averageTimeToPostFromFirst = timeDiffsFromFirst.length > 0
-    ? Math.round((timeDiffsFromFirst.reduce((acc, curr) => acc + curr, 0) / timeDiffsFromFirst.length) * 10) / 10
-    : 0
-    
-  return {
-    numberOfPosts,
-    medianWordCount,
-    averageTimeDiff,
-    firstReplyTimestamp,
-    averageTimeToPostFromFirst,
-    averagePostsPerAuthor
-  }
 }
 
 // Function to write the summary to CSV
